# docs/swagger.yaml
definitions:
  api.CheckStatusResp:
    properties:
      status:
        type: boolean
    type: object
  api.Resp:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
  api.SetStatusReq:
    properties:
      password:
        type: string
      programs_name:
        type: string
      status:
        type: boolean
      username:
        type: string
    type: object
info:
  contact: {}
paths:
  /change/checkStatus:
    post:
      consumes:
      - application/json
      description: 根据项目名称检查该变更项目的状态
      parameters:
      - description: 项目名称
        in: query
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            allOf:
            - $ref: '#/definitions/api.Resp'
            - properties:
                data:
                  $ref: '#/definitions/api.CheckStatusResp'
              type: object
        "400":
          description: 错误响应
          schema:
            $ref: '#/definitions/api.Resp'
      summary: 检查变更项目状态
      tags:
      - 变更项目状态管理
  /change/setStatus:
    put:
      consumes:
      - application/json
      description: 根据项目名称和状态更新该变更项目的状态
      parameters:
      - description: 项目状态更新请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.SetStatusReq'
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            $ref: '#/definitions/api.Resp'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/api.Resp'
        "401":
          description: 认证失败
          schema:
            $ref: '#/definitions/api.Resp'
        "500":
          description: 保存失败
          schema:
            $ref: '#/definitions/api.Resp'
      summary: 设置变更项目状态
      tags:
      - 变更项目状态管理
  /checkStatus:
    post:
      consumes:
      - application/json
      description: 根据项目名称检查该项目的状态
      parameters:
      - description: 项目名称
        in: query
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            allOf:
            - $ref: '#/definitions/api.Resp'
            - properties:
                data:
                  $ref: '#/definitions/api.CheckStatusResp'
              type: object
        "400":
          description: 错误响应
          schema:
            $ref: '#/definitions/api.Resp'
      summary: 检查项目状态
      tags:
      - 项目状态管理
  /setStatus:
    put:
      consumes:
      - application/json
      description: 根据项目名称和状态更新该项目的状态
      parameters:
      - description: 项目状态更新请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.SetStatusReq'
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            $ref: '#/definitions/api.Resp'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/api.Resp'
        "401":
          description: 认证失败
          schema:
            $ref: '#/definitions/api.Resp'
        "500":
          description: 保存失败
          schema:
            $ref: '#/definitions/api.Resp'
      summary: 设置项目状态
swagger: "2.0"